+++ CPanel/particleFMM.cpp
@@ -23,7 +23,7 @@ void particleFMM::computeMultExp(){
     { // Ignore root node. 1 based indexing for tree
         std::vector<node<particle> *> nodes = partTree->getLevelNodes(i);
         
-        for (int j=0; j<nodes.size(); j++)
+        for (size_t j=0; j<nodes.size(); j++)
         {
             if (nodes[j]->isLeafNode())
             {
@@ -59,7 +59,7 @@ Eigen::Vector3d particleFMM::barnesHutVel(Eigen::Vector3d POI){
     Eigen::Vector3d velInfl = Eigen::Vector3d::Zero();
     std::vector<node<particle>*> rootChildren = partTree->getRootNode()->getChildren();
     
-    for (int i=0; i<rootChildren.size(); i++)
+    for (size_t i=0; i<rootChildren.size(); i++)
     {
         velInfl+=rootChildren[i]->calcVel(POI);
     }
@@ -74,7 +74,7 @@ Eigen::Vector3d particleFMM::barnesHutVel(particle* part){
     Eigen::Vector3d velInfl = Eigen::Vector3d::Zero();
     std::vector<node<particle>*> rootChildren = partTree->getRootNode()->getChildren();
     
-    for (int i=0; i<rootChildren.size(); i++)
+    for (size_t i=0; i<rootChildren.size(); i++)
     {
         velInfl += rootChildren[i]->calcVel(part);
     }
@@ -86,7 +86,7 @@ Eigen::Vector3d particleFMM::barnesHutStretch(particle* part){
     Eigen::Vector3d stretchInfl = Eigen::Vector3d::Zero();
     std::vector<node<particle>*> rootChildren = partTree->getRootNode()->getChildren();
     
-    for (int i=0; i<rootChildren.size(); i++)
+    for (size_t i=0; i<rootChildren.size(); i++)
     {
         stretchInfl += rootChildren[i]->calcStretch(part);
     }
@@ -98,7 +98,7 @@ Eigen::Vector3d particleFMM::barnesHutDiff(particle* part){
     Eigen::Vector3d diffInfl = Eigen::Vector3d::Zero();
     std::vector<node<particle>*> rootChildren = partTree->getRootNode()->getChildren();
     
-    for (int i=0; i<rootChildren.size(); i++)
+    for (size_t i=0; i<rootChildren.size(); i++)
     {
         diffInfl += rootChildren[i]->calcDiff(part);
     }
@@ -118,7 +118,7 @@ Eigen::Vector3d particleFMM::findExpPos(node<particle>* thisNode)
     Eigen::Vector3d weightedPos = Eigen::Vector3d::Zero();
     double totStrength = 0;
     
-    for (int i=0; i<nParts.size(); i++) {
+    for (size_t i=0; i<nParts.size(); i++) {
         weightedPos += (nParts[i]->pos) * (nParts[i]->strength).norm();
         totStrength += (nParts[i]->strength).norm();
     }
@@ -133,7 +133,7 @@ Eigen::Vector3d particleFMM::findExpPos(std::vector<particle*> parts)
     Eigen::Vector3d weightedPos = Eigen::Vector3d::Zero();
     double totStrength = 0;
     
-    for (int i=0; i<parts.size(); i++) {
+    for (size_t i=0; i<parts.size(); i++) {
         weightedPos += (parts[i]->pos) * (parts[i]->strength).norm();
         totStrength += (parts[i]->strength).norm();
     }
@@ -147,7 +147,7 @@ Eigen::Vector3d particleFMM::findExpStrength(node<particle>* thisNode){
     
     Eigen::Vector3d nodeStren = Eigen::Vector3d::Zero();
     
-    for (int i=0; i<nodeParticles.size(); i++) {
+    for (size_t i=0; i<nodeParticles.size(); i++) {
         nodeStren+=nodeParticles[i]->strength;
     }
     
@@ -158,7 +158,7 @@ Eigen::Vector3d particleFMM::findExpStrength(std::vector<particle*> parts){
     
     Eigen::Vector3d nodeStren = Eigen::Vector3d::Zero();
     
-    for (int i=0; i<parts.size(); i++) {
+    for (size_t i=0; i<parts.size(); i++) {
         nodeStren+=parts[i]->strength;
     }
     
@@ -171,7 +171,7 @@ double particleFMM::findExpRadius(node<particle>* thisNode){
     
     double nodeRad = 0;
     
-    for (int i=0; i<nodeParticles.size(); i++) {
+    for (size_t i=0; i<nodeParticles.size(); i++) {
         nodeRad+=nodeParticles[i]->radius;
     }
     
@@ -183,7 +183,7 @@ double particleFMM::findExpRadius(std::vector<particle*> parts){
     
     double nodeRad = 0;
     
-    for (int i=0; i<parts.size(); i++) {
+    for (size_t i=0; i<parts.size(); i++) {
         nodeRad+=parts[i]->radius;
     }
